refactor(🚚): moving store in a folder strcuture
